+++ packages/watcher-data-extractor/tests/extractor/permitExtractor.spec.ts
@@ -23,7 +23,7 @@ import {
 import { JsonBI } from '../../lib/utils';
 import { ExtractedPermit } from '../../lib/interfaces/extractedPermit';
 
-jest.mock('@rosen-clients/ergo-explorer');
+vi.mock('@rosen-clients/ergo-explorer');
 let dataSource: DataSource;
 let repository: Repository<PermitEntity>;
 
@@ -187,7 +187,7 @@ describe('permitExtractor', () => {
      * - should extract permit data from api output
      */
     it('should extract permit data from api output', async () => {
-      jest.mocked(ergoExplorerClientFactory).mockReturnValue({
+      vi.mocked(ergoExplorerClientFactory).mockReturnValue({
         v1: {
           getApiV1TransactionsP1: async () => ({
             blockId: 'blockId',
@@ -232,7 +232,7 @@ describe('permitExtractor', () => {
      * - should extract permit data from api output
      */
     it('should extract block id and height for a transaction', async () => {
-      jest.mocked(ergoExplorerClientFactory).mockReturnValue({
+      vi.mocked(ergoExplorerClientFactory).mockReturnValue({
         v1: {
           getApiV1TransactionsP1: async () => ({
             blockId: 'blockId',
@@ -268,7 +268,7 @@ describe('permitExtractor', () => {
      * - should get data and extract information twice when data count is more than api limit
      */
     it('should iterate on api when data count is more than api limit', async () => {
-      jest.mocked(ergoExplorerClientFactory).mockReturnValue({
+      vi.mocked(ergoExplorerClientFactory).mockReturnValue({
         v1: {
           getApiV1BoxesUnspentByergotreeP1: async () => ({
             items: [],
@@ -283,7 +283,7 @@ describe('permitExtractor', () => {
         'RWT',
         'url',
       );
-      const spy = jest
+      const spy = vi
         .spyOn(extractor, 'extractPermitData')
         .mockResolvedValue([sampleExtractedPermit]);
       const result = await extractor.getAllUnspentPermits(100);
@@ -310,7 +310,7 @@ describe('permitExtractor', () => {
           creationHeight: 120,
         },
       ];
-      jest.mocked(ergoExplorerClientFactory).mockReturnValue({
+      vi.mocked(ergoExplorerClientFactory).mockReturnValue({
         v1: {
           getApiV1BoxesUnspentByergotreeP1: async () => ({
             items: boxes,
@@ -325,7 +325,7 @@ describe('permitExtractor', () => {
         'RWT',
         'url',
       );
-      const spy = jest
+      const spy = vi
         .spyOn(extractor, 'extractPermitData')
         .mockResolvedValue([]);
       await extractor.getAllUnspentPermits(110);
@@ -350,7 +350,7 @@ describe('permitExtractor', () => {
       const box = {
         boxId: 'boxId',
       };
-      jest.mocked(ergoExplorerClientFactory).mockReturnValue({
+      vi.mocked(ergoExplorerClientFactory).mockReturnValue({
         v1: {
           getApiV1BoxesP1: async () => box,
         },
@@ -362,7 +362,7 @@ describe('permitExtractor', () => {
         'RWT',
         'url',
       );
-      const spy = jest
+      const spy = vi
         .spyOn(extractor, 'extractPermitData')
         .mockResolvedValue([sampleExtractedPermit]);
       const boxData = await extractor.getPermitWithBoxId('boxId');
@@ -397,7 +397,7 @@ describe('permitExtractor', () => {
      */
     it('should remove invalid permit from database', async () => {
       await insertPermitEntity(dataSource, 'boxId');
-      const spy = jest
+      const spy = vi
         .spyOn(extractor, 'getPermitWithBoxId')
         .mockResolvedValue(undefined);
 
@@ -420,12 +420,10 @@ describe('permitExtractor', () => {
      */
     it('should update valid box information when spent bellow the initial height', async () => {
       await insertPermitEntity(dataSource, 'boxId');
-      const spy = jest
-        .spyOn(extractor, 'getPermitWithBoxId')
-        .mockResolvedValue({
-          spendBlock: 'spendBlockId',
-          spendHeight: 99,
-        } as any);
+      const spy = vi.spyOn(extractor, 'getPermitWithBoxId').mockResolvedValue({
+        spendBlock: 'spendBlockId',
+        spendHeight: 99,
+      } as any);
 
       await extractor.validateOldStoredPermits(['boxId'], 100);
       expect(spy).toHaveBeenCalledWith('boxId');
@@ -448,12 +446,10 @@ describe('permitExtractor', () => {
      */
     it('should not change valid permit information when spent after the initial height', async () => {
       await insertPermitEntity(dataSource, 'boxId');
-      const spy = jest
-        .spyOn(extractor, 'getPermitWithBoxId')
-        .mockResolvedValue({
-          spendBlock: 'spendBlockId',
-          spendHeight: 120,
-        } as any);
+      const spy = vi.spyOn(extractor, 'getPermitWithBoxId').mockResolvedValue({
+        spendBlock: 'spendBlockId',
+        spendHeight: 120,
+      } as any);
 
       await extractor.validateOldStoredPermits(['boxId'], 100);
       expect(spy).toHaveBeenCalledWith('boxId');
@@ -500,10 +496,10 @@ describe('permitExtractor', () => {
         height: 99,
         WID: 'wid2',
       };
-      jest
-        .spyOn(extractor, 'getAllUnspentPermits')
-        .mockResolvedValue([extractedPermit]);
-      jest.spyOn(extractor, 'validateOldStoredPermits').mockImplementation();
+      vi.spyOn(extractor, 'getAllUnspentPermits').mockResolvedValue([
+        extractedPermit,
+      ]);
+      vi.spyOn(extractor, 'validateOldStoredPermits').mockImplementation();
       await extractor.initializeBoxes({ height: 100 } as BlockInfo);
       const permit = await repository.findOne({ where: { boxId: 'boxId2' } });
       expect(permit).not.toBeNull();
@@ -538,10 +534,10 @@ describe('permitExtractor', () => {
         height: 100,
         WID: 'wid',
       };
-      jest
-        .spyOn(extractor, 'getAllUnspentPermits')
-        .mockResolvedValue([extractedPermit]);
-      jest.spyOn(extractor, 'validateOldStoredPermits').mockImplementation();
+      vi.spyOn(extractor, 'getAllUnspentPermits').mockResolvedValue([
+        extractedPermit,
+      ]);
+      vi.spyOn(extractor, 'validateOldStoredPermits').mockImplementation();
       await insertPermitEntity(dataSource, 'boxId1');
       await extractor.initializeBoxes({ height: 100 } as BlockInfo);
       const permit = await repository.findOne({ where: { boxId: 'boxId1' } });