+++ packages/scanner-interfaces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rosen-bridge/scanner-interfaces
 
+## 0.2.0
+
+### Minor Changes
+
+- Upgrade to Node.js version 22.18.0
+
 ## 0.1.1
 
 ### Patch Changes